renamed statewide_test file and restructured test to use real data
