+++ lib/statewide_test_repository.rb
@@ -1,5 +1,6 @@
 require 'pry'
 require_relative '../lib/statewide_test'
+require_relative '../lib/data_manager'
 
 class StatewideTestRepository
   attr_reader :stw_tests
